feat: addition of a new stylization of yup errors and a visual improvement and code ordering :sparkles:
A new stylization of yup errors was added, aligning the TextFiled to the left and visually improving the refactoring of the v-dialog and a better visual organization of the folders in addition to the independence of the files.
