--- components/Pages/Login/Login.vue
@@ -1,7 +1,0 @@
-<script lang="ts" setup>
-useTitle(null, 'Login');
-</script>
-
-<template>
-	<PagesLoginContent />
-</template>
