+++ components/Section/Ambientes/Grid/Grid.vue
@@ -19,7 +19,7 @@ const { ambientes } = await useApiAmbientesFindAll(searchBarText);
 			>
 				<template #title>
 					{{ ambiente.nome }}
-				</template>
+				</template>	
 
 				<template #actions>
           <SectionAmbientesModal :editId="ambiente.id" />
